requirements.txt
infi.systray
setuptools
py2exe